refactoring and showing spectrum
