+++ web/src/analyzer/AnalyzerNeedsAccess.js
@@ -1,0 +1,8 @@
+import React from 'react';
+
+
+export function AnalyzerNeedsAccess() {
+    return (
+        <h1>analyzer needs access</h1>
+    );
+}
